Add listener file to state package
